+++ src/amqp_management.pyx
@@ -15,10 +15,10 @@ cimport c_message
 _logger = logging.getLogger(__name__)
 
 
-cpdef create_management_operation(cSession session, management_node):
+cpdef create_management_operation(cSession session, management_node, encoding):
     mgr_op = cManagementOperation()
     if isinstance(management_node, str):
-        management_node = management_node.encode('utf-8')
+        management_node = management_node.encode(encoding)
     mgr_op.create(session, <const char*>management_node)
     return mgr_op