+++ uamqp/async/connection_async.py
@@ -68,6 +68,7 @@ class ConnectionAsync(connection.Connection):
                  properties=None,
                  remote_idle_timeout_empty_frame_send_ratio=None,
                  debug=False,
+                 encoding='UTF-8'
                  loop=None):
         self.loop = loop or asyncio.get_event_loop()
         super(ConnectionAsync, self).__init__(
@@ -78,7 +79,8 @@ class ConnectionAsync(connection.Connection):
             idle_timeout=idle_timeout,
             properties=properties,
             remote_idle_timeout_empty_frame_send_ratio=remote_idle_timeout_empty_frame_send_ratio,
-            debug=debug)
+            debug=debug,
+            encoding=encoding)
         self._lock = asyncio.Lock(loop=self.loop)
 
     async def __aenter__(self):